Add files via upload
This case study focuses on the creation of a sales report using Power BI. The data has been extracted from various sources and transformed according to the specified requirements. The goal was to assemble a comprehensive sales report dashboard that provides valuable insights.

To consolidate the data, a mechanism was implemented to load all the files from the Sales folder into a single Sales fact table. This mechanism was designed to be resilient, ensuring that the absence of files or the addition of new yearly sales files would not cause errors upon refreshing the data.

Data modeling was performed to enhance the structure and usability of the data. The Sales fact table underwent several transformations, including the splitting of the "Location" field into separate "Country" and "City" fields, allowing for geospatial analysis. The Date field was also modified to ensure proper formatting.

Unique keys (GeoKey) were created in both the Sales and Geography tables to establish relationships and enable efficient data analysis. Additionally, a small function was developed to clean the ID columns in the SalesRep and SubCategory queries, removing the "ID - " prefix for consistency.


DAX calculations were applied to derive important metrics for analysis. These calculations included determining the Total Revenue by multiplying the Product's Retail Price with the Units, calculating the Total Cost by multiplying the Product's Standard Cost with the Units, and evaluating the Gross Profit as the difference between Total Revenue and Total Cost. The Gross Profit MoM growth Change% measure was calculated to provide insights for decision-making. The Average Sales per day measure was derived by summing the Total Revenue per day based on the actual sales dates.

To present the sales insights effectively, a powerful dashboard was created using various visuals. The visuals were designed to showcase the sales performance, trends, and breakdown analysis by product. When plotting the Month on the x-axis, the months were sorted from January to December to ensure a consistent representation.

Overall, this case study demonstrates the successful extraction and transformation of data using Power BI, resulting in a powerful sales report dashboard that provides valuable insights for effective decision-making.